refactor(apps): adopt feature-first module layout
