+++ app/feiwen/src/foundation.rs
@@ -1,0 +1,3 @@
+pub(crate) mod i18n;
+
+pub(crate) use i18n::I18n;
