+++ app/http-client/src/features.rs
@@ -1,0 +1,3 @@
+pub(crate) mod request;
+
+pub(crate) use request::HttpFormView;
