+++ app/novel-download/src/features.rs
@@ -1,0 +1,3 @@
+mod workspace;
+
+pub(crate) use workspace::WorkspaceView;
