--- app/novel-download/src/views.rs
@@ -1,2 +1,0 @@
-mod worksapce;
-pub use worksapce::WorkspaceView;
